+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/rrteamcode/messages/TankCommandMessage.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.teamcode.messages;
+package org.firstinspires.ftc.teamcode.rrteamcode.messages;
 
 public final class TankCommandMessage {
     public long timestamp;
