+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/rrteamcode/messages/ThreeDeadWheelInputsMessage.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.teamcode.messages;
+package org.firstinspires.ftc.teamcode.rrteamcode.messages;
 
 import com.acmerobotics.roadrunner.ftc.PositionVelocityPair;
