+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/rrteamcode/tuning/SplineTest.java
@@ -1,12 +1,12 @@
-package org.firstinspires.ftc.teamcode.teamcode.tuning;
+package org.firstinspires.ftc.teamcode.rrteamcode.tuning;
 
 import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
-import org.firstinspires.ftc.teamcode.MecanumDrive;
-import org.firstinspires.ftc.teamcode.TankDrive;
+import org.firstinspires.ftc.teamcode.rrteamcode.MecanumDrive;
+import org.firstinspires.ftc.teamcode.rrteamcode.TankDrive;
 
 public final class SplineTest extends LinearOpMode {
     @Override
