+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/rrteamcode/tuning/TuningOpModes.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.teamcode.tuning;
+package org.firstinspires.ftc.teamcode.rrteamcode.tuning;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.config.reflection.ReflectionConfig;
@@ -22,10 +22,10 @@ import com.qualcomm.robotcore.eventloop.opmode.OpModeManager;
 import com.qualcomm.robotcore.eventloop.opmode.OpModeRegistrar;
 
 import org.firstinspires.ftc.robotcore.internal.opmode.OpModeMeta;
-import org.firstinspires.ftc.teamcode.MecanumDrive;
-import org.firstinspires.ftc.teamcode.TankDrive;
-import org.firstinspires.ftc.teamcode.ThreeDeadWheelLocalizer;
-import org.firstinspires.ftc.teamcode.TwoDeadWheelLocalizer;
+import org.firstinspires.ftc.teamcode.rrteamcode.MecanumDrive;
+import org.firstinspires.ftc.teamcode.rrteamcode.TankDrive;
+import org.firstinspires.ftc.teamcode.rrteamcode.ThreeDeadWheelLocalizer;
+import org.firstinspires.ftc.teamcode.rrteamcode.TwoDeadWheelLocalizer;
 
 import java.util.ArrayList;
 import java.util.Arrays;
